refactor: 도메인에서 자원별로 service 분리
