+++ src/main/java/com/umc/product/curriculum/application/port/in/query/GetChallengerWorkbookUseCase.java
@@ -12,6 +12,9 @@ import java.util.List;
  */
 public interface GetChallengerWorkbookUseCase {
 
+    // <----------------------------- 챌린저 기능 ----------------------------->
+
+    // <----------------------------- 운영진 기능 ----------------------------->
     /**
      * (운영진 기능) 학교, 주차, 스터디그룹별 제출된 ChallengerWorkbook 리스트 조회
      *
